fix: Refactored server logic
