+++ .gitignore
@@ -1,1 +1,7 @@
 .dist/
+venv/
+__pycache__/
+
+# userauth db
+server/instance/
+instance/site.db
