--- server/.gitignore
@@ -1,3 +1,0 @@
-venv/
-instance/
-/__pycache__/
