+++ server/api/__init__.py
@@ -1,0 +1,1 @@
+# api/__init__.py
