+++ server/config.py
@@ -1,0 +1,3 @@
+class Config:
+    SECRET_KEY = 'your_secret_key'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///site.db'
