+++ server/extensions.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+from flask_cors import CORS
+
+db = SQLAlchemy()
+bcrypt = Bcrypt()
+cors = CORS()
